--- a/src/io/parquet/write/mod.rs
+++ b/src/io/parquet/write/mod.rs
@@ -143,8 +143,9 @@ pub fn array_to_pages(
     // we also check for an array.len > 3 to prevent infinite recursion
     // still have to figure out how to deal with values that are i32::MAX size, such as very large
     // strings or a list column with many elements
-
-    if (estimated_bytes_size(array)) >= (2u32.pow(31) - 2u32.pow(25)) as usize && array.len() > 3 {
+    
+    let array_byte_size = estimated_bytes_size(array.as_ref());
+    if array_byte_size >= (2u32.pow(31) - 2u32.pow(25)) as usize && array.len() > 3 {
         let split_at = array.len() / 2;
         let left = array.slice(0, split_at);
         let right = array.slice(split_at, array.len() - split_at);
@@ -169,27 +170,21 @@ pub fn array_to_pages(
             _ => {
                 const DEFAULT_PAGE_SIZE: usize = 1024 * 1024;
                 let page_size = options.data_pagesize_limit.unwrap_or(DEFAULT_PAGE_SIZE);
-
-                let bytes_per_row = estimated_bytes_size(array.slice(0, 1).as_ref()).max(1);
-                let row_per_page = page_size / bytes_per_row + 1;
-
-                if array.len() <= row_per_page {
-                    array_to_page(array, type_, nested, options, encoding)
-                        .map(|page| DynIter::new(std::iter::once(Ok(page))))
-                } else {
-                    let first = array.slice(0, row_per_page);
-                    let other = array.slice(row_per_page, array.len() - row_per_page);
-                    Ok(DynIter::new(
-                        array_to_pages(first.as_ref(), type_.clone(), nested, options, encoding)?
-                            .chain(array_to_pages(
-                                other.as_ref(),
-                                type_,
-                                nested,
-                                options,
-                                encoding,
-                            )?),
-                    ))
-                }
+                let bytes_per_row = ((array_byte_size as f64) / (array.len() as f64)) as usize;
+                let row_per_page = (page_size / bytes_per_row).max(1);
+                
+                let iter = (0..array.len()).step_by(row_per_page).map(move |offset| {
+                    let length = if offset + row_per_page > array.len() {
+                        array.len() - offset
+                    } else {
+                        row_per_page
+                    };
+                    
+                    let sub_array = array.slice(offset, length);
+                    array_to_page(sub_array.as_ref(), type_, nested, options, encoding)
+                });
+                
+                Ok(DynIter::new(iter))
             }
         }
     }