--- a/src/io/parquet/write/mod.rs
+++ b/src/io/parquet/write/mod.rs
@@ -143,7 +143,7 @@ pub fn array_to_pages(
     // we also check for an array.len > 3 to prevent infinite recursion
     // still have to figure out how to deal with values that are i32::MAX size, such as very large
     // strings or a list column with many elements
-    
+
     let array_byte_size = estimated_bytes_size(array.as_ref());
     if array_byte_size >= (2u32.pow(31) - 2u32.pow(25)) as usize && array.len() > 3 {
         let split_at = array.len() / 2;
@@ -171,20 +171,23 @@ pub fn array_to_pages(
                 const DEFAULT_PAGE_SIZE: usize = 1024 * 1024;
                 let page_size = options.data_pagesize_limit.unwrap_or(DEFAULT_PAGE_SIZE);
                 let bytes_per_row = ((array_byte_size as f64) / (array.len() as f64)) as usize;
-                let row_per_page = (page_size / bytes_per_row).max(1);
-                
-                let iter = (0..array.len()).step_by(row_per_page).map(move |offset| {
-                    let length = if offset + row_per_page > array.len() {
-                        array.len() - offset
-                    } else {
-                        row_per_page
-                    };
-                    
-                    let sub_array = array.slice(offset, length);
-                    array_to_page(sub_array.as_ref(), type_, nested, options, encoding)
-                });
-                
-                Ok(DynIter::new(iter))
+                let rows_per_page = (page_size / bytes_per_row).max(1);
+
+                let vs: Vec<Result<EncodedPage>> = (0..array.len())
+                    .step_by(rows_per_page)
+                    .map(|offset| {
+                        let length = if offset + rows_per_page > array.len() {
+                            array.len() - offset
+                        } else {
+                            rows_per_page
+                        };
+
+                        let sub_array = array.slice(offset, length);
+                        array_to_page(sub_array.as_ref(), type_.clone(), nested, options, encoding)
+                    })
+                    .collect();
+
+                Ok(DynIter::new(vs.into_iter()))
             }
         }
     }