--- a/src/io/parquet/write/sink.rs
+++ b/src/io/parquet/write/sink.rs
@@ -35,6 +35,7 @@ use super::{Encoding, SchemaDescriptor, WriteOptions};
 ///     write_statistics: true,
 ///     compression: CompressionOptions::Uncompressed,
 ///     version: Version::V2,
+///     data_pagesize_limit: None,
 /// };
 ///
 /// let mut buffer = vec![];