arrow2 writes single page by default which will hurt the read performance a lot (up to 4x slower than the original file).

This pr introduces an option named `data_pagesize_limit`  to split large Array into small pages.

fixes #1291