@sundy-li 2022-11-23
Read write benchmark gists:  https://gist.github.com/sundy-li/4984ec7cfeade556d60306a3a218ec8a


We use TPCH's  lineitem table, it's original file is generated by dremio, it's well paged for column `l_quantity`

```
column ["l_quantity"]: num_of_pages 460, num_of_rows 5192615, size 10454182, size(uncompressed) 41555178
```

We use [parquet_write.rs](https://gist.github.com/sundy-li/4984ec7cfeade556d60306a3a218ec8a#file-parquet_write-rs) to generate the output file by input file and  [parquet_read.rs](https://gist.github.com/sundy-li/4984ec7cfeade556d60306a3a218ec8a#file-parquet_read-rs) to test the performance read on parallel.


Before (single large page per rowgroup):

```
/tmp/input.parquet, parallel 1 read took: 21 ms
/tmp/input.parquet, parallel 2 read took: 21 ms
/tmp/input.parquet, parallel 4 read took: 23 ms
/tmp/input.parquet, parallel 8 read took: 37 ms
/tmp/input.parquet, parallel 16 read took: 53 ms
/tmp/output.parquet, parallel 1 read took: 37 ms
/tmp/output.parquet, parallel 2 read took: 42 ms
/tmp/output.parquet, parallel 4 read took: 65 ms
/tmp/output.parquet, parallel 8 read took: 124 ms
/tmp/output.parquet, parallel 16 read took: 284 ms
```

As you can see, it's 5X worse in reading on 16 parallelism.

After this pr:

```
/tmp/input.parquet, parallel 1 read took: 21 ms
/tmp/input.parquet, parallel 2 read took: 22 ms
/tmp/input.parquet, parallel 4 read took: 24 ms
/tmp/input.parquet, parallel 8 read took: 29 ms
/tmp/input.parquet, parallel 16 read took: 52 ms
/tmp/output.parquet, parallel 1 read took: 23 ms
/tmp/output.parquet, parallel 2 read took: 25 ms
/tmp/output.parquet, parallel 4 read took: 27 ms
/tmp/output.parquet, parallel 8 read took: 31 ms
/tmp/output.parquet, parallel 16 read took: 48 ms
```