@jorgecarleitao 2022-11-23
Out of curiosity, what is the difference in file size? Usually smaller pages => larger files.