@sundy-li 2022-11-23
> There is a similar limit in arrow-rs

Thank you @alamb,  I can't figure out a better name so I keep it compatible with arrow-rs.  The row count limit is less useful than the bytes limit, I'll try to add it in the next pr.

I found this best practice to optimize parquet files:  https://docs.dremio.com/software/data-formats/parquet-files/

The page-size defaults to ~100KB seem to be really reasonable. (I have test from 8192byte to 1MB today and found it has the best performance in this default config)